+++ src/main/java/com/example/backend/cart/dto/CartItemResponseDTO.java
@@ -1,5 +1,7 @@
 package com.example.backend.cart.dto;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -14,6 +16,6 @@ public class CartItemResponseDTO {
     private Long cartId;
     private Long productId;
     private Integer quantity;
-    private Integer unitPrice;
-    private Integer totalPrice;
+    private BigDecimal unitPrice;
+    private BigDecimal totalPrice;
 }
